feat: 增加优化启停脚本

+++ build/stop_all.sh
@@ -1,5 +1,11 @@
-#!/bin/sh
+#!/usr/bin/bash
 
 pids=$( ps -ef | grep java | grep cloudtimes | grep -v grep | awk '{print $2}' )
-echo kill $pids
-kill -9 $pids
+if [[ -n "$pids" ]]; then
+  echo kill $pids
+  kill -9 $pids
+else
+  echo "no cloudtimes services is running"
+fi
+
+exit 0
